refactor: isolate hypothesis classification
